Fix sample app package naming
