+++ app/src/main/AndroidManifest.xml
@@ -12,26 +12,26 @@
 
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.xfinity.blueprint_sample">
+    package="com.xfinity.blueprint.sample">
 
     <application
-        android:name=".BlueprintSampleApplication"
+        android:name="com.xfinity.blueprint.sample.BlueprintSampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.xfinity.blueprint_sample.LauncherActivity">
+        <activity android:name="com.xfinity.blueprint.sample.LauncherActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name="com.xfinity.blueprint_sample.StaticScreenActivity" android:label="@string/static_screen"/>
-        <activity android:name="com.xfinity.blueprint_sample.DynamicScreenActivity" android:label="@string/dynamic_screen"/>
-        <activity android:name="com.xfinity.blueprint_sample.ArchitectSampleActivity" android:label="@string/bp_arch_screen"/>
+        <activity android:name="com.xfinity.blueprint.sample.StaticScreenActivity" android:label="@string/static_screen"/>
+        <activity android:name="com.xfinity.blueprint.sample.DynamicScreenActivity" android:label="@string/dynamic_screen"/>
+        <activity android:name="com.xfinity.blueprint.sample.ArchitectSampleActivity" android:label="@string/bp_arch_screen"/>
     </application>
 
 </manifest>
