+++ app/src/main/java/com/xfinity/blueprint/sample/mvp/model/DynamicScreenModel.kt
@@ -9,11 +9,11 @@
  * limitations under the License.
  */
 
-package com.xfinity.blueprint_sample.mvp.model
+package com.xfinity.blueprint.sample.mvp.model
 
-import com.xfinity.blueprint.model.ComponentModel
-
-class FooterModel : ComponentModel {
-    val footer : String = "This is the footer"
-    var enabled = false
+open class DynamicScreenModel {
+    open var headerModel: HeaderModel = HeaderModel()
+    var footerModel: FooterModel = FooterModel()
+    open var dataItemModels: MutableList<DataItemModel> = mutableListOf(DataItemModel(), DataItemModel(), DataItemModel(),
+            DataItemModel(), DataItemModel(), DataItemModel())
 }
