+++ app/src/main/java/com/xfinity/blueprint/sample/mvp/presenter/DefaultDataItemPresenter.kt
@@ -9,14 +9,14 @@
  * limitations under the License.
  */
 
-package com.xfinity.blueprint_sample.mvp.presenter
+package com.xfinity.blueprint.sample.mvp.presenter
 
 import com.xfinity.blueprint.event.ComponentEventManager
 import com.xfinity.blueprint.presenter.EventEmittingComponentPresenter
+import com.xfinity.blueprint.sample.mvp.model.DataItemModel
+import com.xfinity.blueprint.sample.mvp.view.DataItemView
 import com.xfinity.blueprint_annotations.DefaultPresenter
 import com.xfinity.blueprint_annotations.DefaultPresenterConstructor
-import com.xfinity.blueprint_sample.mvp.model.DataItemModel
-import com.xfinity.blueprint_sample.mvp.view.DataItemView
 
 @DefaultPresenter(viewClass = DataItemView::class)
 class DefaultDataItemPresenter
