+++ app/src/main/java/com/xfinity/blueprint/sample/mvp/presenter/DynamicScreenPresenter.kt
@@ -9,18 +9,18 @@
  * limitations under the License.
  */
 
-package com.xfinity.blueprint_sample.mvp.presenter
+package com.xfinity.blueprint.sample.mvp.presenter
 
+import com.xfinity.blueprint.blueprint.AppComponentRegistry
 import com.xfinity.blueprint.event.ComponentEvent
 import com.xfinity.blueprint.event.ComponentEventManager
 import com.xfinity.blueprint.model.Component
 import com.xfinity.blueprint.model.ComponentModel
 import com.xfinity.blueprint.presenter.DefaultComponentPresenter
 import com.xfinity.blueprint.presenter.EventHandlingScreenPresenter
-import com.xfinity.blueprint_sample.blueprint.AppComponentRegistry
-import com.xfinity.blueprint_sample.mvp.model.DataItemModel
-import com.xfinity.blueprint_sample.mvp.model.DynamicScreenModel
-import com.xfinity.blueprint_sample.mvp.view.DefaultDynamicScreenView
+import com.xfinity.blueprint.sample.mvp.model.DataItemModel
+import com.xfinity.blueprint.sample.mvp.model.DynamicScreenModel
+import com.xfinity.blueprint.sample.mvp.view.DefaultDynamicScreenView
 import java.util.Timer
 import java.util.TimerTask
