+++ app/src/main/java/com/xfinity/blueprint/sample/mvp/view/StaticScreenView.kt
@@ -9,19 +9,10 @@
  * limitations under the License.
  */
 
-package com.xfinity.blueprint_sample.mvp.view
+package com.xfinity.blueprint.sample.mvp.view
 
-import android.support.v7.widget.RecyclerView
-import android.widget.TextView
-import com.xfinity.blueprint_annotations.ComponentViewClass
-import com.xfinity.blueprint_annotations.ComponentViewHolder
-import com.xfinity.blueprint_sample.R
+import com.xfinity.blueprint.view.ScreenView
+import com.xfinity.blueprint.view.ScreenViewDelegate
 
-@ComponentViewClass(viewHolderClass = FooterViewHolder::class)
-class FooterView : FooterViewBase()
-
-@ComponentViewHolder(viewType = R.layout.footer_view)
-class FooterViewHolder(itemView: android.view.View) : RecyclerView.ViewHolder(itemView) {
-    val footer : TextView = itemView.findViewById(R.id.footer) as TextView
-}
-
+class StaticScreenView(val screenViewDelegate: ScreenViewDelegate) :
+        ScreenView by screenViewDelegate
